企业管理----企业员工管理

+++ xunyu_department/src/main/java/com/xunyu/xunyu_department/controller/UserController.java
@@ -46,11 +46,10 @@ public class UserController {
     @RequestMapping(value = "/addemployee",method = RequestMethod.POST)
     @ResponseBody
     public Result<String> addEmployee(UsersModel model){
-//        Result result = checkLogin(new Result(),model.getSessionId());
-//        if (result.getCode() != null){
-//            return result;
-//        }
-        Result result = new Result();
+        Result result = checkLogin(new Result(),model.getSessionId());
+        if (result.getCode() != null){
+            return result;
+        }
         try{
             List<Users> users = userService.selectUserListByAccount(model);
             if (users != null && users.size()>0){
@@ -78,11 +77,10 @@ public class UserController {
     @RequestMapping(value = "/editemployeeinfo",method = RequestMethod.POST)
     @ResponseBody
     public Result<String> editEmployeeInfo(UsersModel model){
-//        Result result = checkLogin(new Result(),model.getSessionId());
-//        if (result.getCode() != null){
-//            return result;
-//        }
-        Result result = new Result();
+        Result result = checkLogin(new Result(),model.getSessionId());
+        if (result.getCode() != null){
+            return result;
+        }
         try {
             if (model.getPassword() != null){
                 model.setPassword(MD5Utils.getMD5(model.getPassword()));
@@ -106,11 +104,10 @@ public class UserController {
     @RequestMapping(value = "/delemployee",method = RequestMethod.POST)
     @ResponseBody
     public Result<String> deleteEmployee(UsersModel model){
-//        Result result = checkLogin(new Result(),model.getSessionId());
-//        if (result.getCode() != null){
-//            return result;
-//        }
-        Result result = new Result();
+        Result result = checkLogin(new Result(),model.getSessionId());
+        if (result.getCode() != null){
+            return result;
+        }
         try {
             int success = userService.deleteUser(model);
             if (success>0){
@@ -131,11 +128,10 @@ public class UserController {
     @RequestMapping(value = "getemployeelist",method = RequestMethod.POST)
     @ResponseBody
     public Result<List<UsersVO>> getEmployeeList(UsersModel model){
-//        Result result = checkLogin(new Result(),model.getSessionId());
-//        if (result.getCode() != null){
-//            return result;
-//        }
-        Result result = new Result();
+        Result result = checkLogin(new Result(),model.getSessionId());
+        if (result.getCode() != null){
+            return result;
+        }
         try {
             List<UsersVO> list = userService.selectUserList(model);
             operationSuccess(result);
@@ -156,11 +152,10 @@ public class UserController {
     @RequestMapping(value = "/batchdelemployee",method = RequestMethod.POST)
     @ResponseBody
     public Result<String> batchDeleteEmployee(UsersModel model){
-//        Result result = checkLogin(new Result(),model.getSessionId());
-//        if (result.getCode() != null){
-//            return result;
-//        }
-        Result result = new Result();
+        Result result = checkLogin(new Result(),model.getSessionId());
+        if (result.getCode() != null){
+            return result;
+        }
         List<Long> list = model.getIdList();
         try{
             int count = userService.batchDeleteUser(list);
